first occurrence of symbol as keyword leaves scope undecided
Fixes long standing problem.  Originally, if a symbol appeared first
in a func as a keyword, it was implicitly declared extern.  After
discussion on user forum with Eric Thiebaut, this convention was
changed with commit 9202a70212 to declare the variable local.  This
broke a number of interpreted functions (e.g.- hcp fixed by commit
06d33427d1), always in cases where the symbol was the name of a
function called after it appeared as a keyword.  The correct behavior
is to leave the symbol scope undecided after a first appearance as a
keyword.