improve Makepkg support for non-Y_HOME installation
Use cases are:
1. Install yorick packages in a subdirectory of ~/yorick when
   you do not have write permission to the public Y_HOME.
2. Maintain a large yorick-steered physics code in a separate
   directory tree from Y_HOME.

Biggest changes are adding install2 and other targets to Makepkg, and
recoding yorick startup sequence to look for Y_HOME.txt file as a
sibling of the executable.  The Y_HOME.txt file contains the path to
be used as Y_HOME, so that except for this file both the public yorick
and the separate TGT=exe executable are separately relocatable.

Added Y_HOME_PKG global variable to yorick, which records directory
tree for current executable, if that is different from Y_HOME.  Also
added appropriate i, i0, and lib subdirectories of Y_HOME_PKG to the
yorick include path, and modified istart_include to find an i-start
and any site-i subdirectories based on Y_HOME_PKG.